services/agent-api/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
PyYAML
anthropic
httpx
mcp
